Replaced all quotes with programmer's quotes.
Quotation marks were what was causing the problem with the FAQ page! Apparently not all quotation marks are created equal and I needed to do a simple find and replace with a "programmer's quotation mark."

The Clone iOS app used the wrong " mark, which broke the whole page. Thanks to ChrisG on Stack Overflow for this fix!